Don't install CRI-O in k8s nodes using containerd 2.0+.
containerd 2.0+ supports setting up pods with user-namespaces, thereby voiding
the need for CRI-O.

Update the sysbox k8s installation script to check if the K8s node it runs has
containerd 2.0, and if so, skip the installation of CRI-O on the node.

Note that due to a bug, containerd versions 2.0.1 -> 2.0.4 don't properly
support user-namespaces. Thus, for K8s nodes with those we still use CRI-O.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>